add raja and camp libraries
Raja is the LLNL c++ api for portable heterogeneous programming, while
camp is the metaprogramming library used as part of the larger raja
framework.  Since it can reasonably be used independently, I split that
out, but raja also uses a copy of it, so if there's a better way to make
one a dependency of the other I'd be happy to do that.